requirements.txt
discord.py
requests
discord_components
